fix: Use StreamPipes client for fetching permissions
